added support for monetary amounts

+++ src/main/java/com/bootcamp/demo/entity/Scooter.java
@@ -4,6 +4,8 @@ package com.bootcamp.demo.entity;
 import com.bootcamp.demo.enums.State;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
@@ -14,7 +16,7 @@ public class Scooter {
 
     private String serial_number;
     private String brand_name;
-    private double acquisition_cost;
+    private BigDecimal acquisition_cost;
     private int production_year;
     private double weight;
     private State state;
